Organizes .js-files like modules
