+++ js/setup.js
@@ -1,0 +1,3 @@
+const PHOTO_COUNT = 25;
+
+export {PHOTO_COUNT};
