+++ js/utilitary.js
@@ -1,0 +1,18 @@
+import {isPositive} from './check.js';
+import {isBigger} from './check.js';
+
+/* Функция возвращает случайное целое число из переданного диапазона включительно.
+Ожидаемый результат: случайное целое число из диапазона "от...до"
+Диапазон может быть только положительный, включая ноль.
+Обработать, как функция должна вести себя, если передать значение «до» меньшее, чем значение «от», или равное ему.
+
+Если проверки пройдены, то вычисляется целое число из диапазона между "от" и "до".
+*/
+function getRandomInteger(lowerNumber, higherNumber) {
+  if (isPositive(lowerNumber) && isPositive(higherNumber) && isBigger(higherNumber, lowerNumber)) {
+    return Math.floor(Math.random() * (Math.floor(higherNumber) - Math.ceil(lowerNumber) + 1)) + Math.ceil(lowerNumber);
+  }
+  throw new Error('Function getRandomInteger - validation error ');
+}
+
+export {getRandomInteger};
